salon/ai-service/requirements.txt
# GlamBook AI Service - Dependencies

# Core
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-dotenv>=1.0.0

# Database
supabase>=2.3.0
asyncpg>=0.29.0

# Eleven Labs (Voice AI)
elevenlabs>=1.0.0

# OpenAI (for agent logic)
openai>=1.6.0
openai-agents>=0.0.5

# Twilio (Telephony)
twilio>=8.10.0

# WebSocket
websockets>=12.0

# HTTP client
httpx>=0.25.0
aiohttp>=3.9.0

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3

# Logging
structlog>=23.2.0
